Store token creator address along with type for uniqueness in the event multiple tokens have the same ISO
Also added additional logic to handle the now longer token type names but displaying only the ISO in the spinners.

+++ app/src/main/java/com/radixdlt/android/data/model/transaction/TransactionEntity.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import kotlinx.android.parcel.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 @Entity(
@@ -18,5 +19,5 @@ data class TransactionEntity(
     val sent: Boolean,
     val dateUnix: Long,
     val tokenClassISO: String,
-    val tokenClassSubUnits: Int
+    val tokenClassSubUnits: BigDecimal
 ) : Parcelable
